+++ Whiskey/Functions/Find-WhiskeyPowerShellModule.ps1
@@ -41,46 +41,123 @@ function Find-WhiskeyPowerShellModule
     Set-StrictMode -Version 'Latest'
     Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
 
-    # If you want to upgrade the PackageManagement and PowerShellGet versions, you must also update:
-    # * appveyor.yml
-    # * PublishPowerShellModule task
-    Import-Module -Name 'PackageManagement' -RequiredVersion '1.4.7' -Global
-    Import-Module -Name 'PowerShellGet' -RequiredVersion '2.2.5' -Global
+    Write-WhiskeyDebug '\Find-WhiskeyPowerShellModule\' -Indent
 
-    Write-WhiskeyDebug -Message ('{0}  {1} ->' -f $Name,$Version)
-    if( $Version )
+    function Import
     {
-        $atVersionString = ' at version {0}' -f $Version
+        [CmdletBinding()]
+        param(
+            [Parameter(Mandatory)]
+            [String] $Name,
 
-        if( -not [Management.Automation.WildcardPattern]::ContainsWildcardCharacters($version) -and [Version]::TryParse($Version,[ref]$null) )
+            [Parameter(Mandatory)]
+            [Version] $MinimumVersion,
+
+            [Parameter(Mandatory)]
+            [Version] $MaximumVersion
+        )
+
+        Write-WhiskeyDebug "Importing module $($Name) for version $($MinimumVersion) - $($MaximumVersion)."
+        $module = Get-Module -Name $Name
+        if( $module )
         {
-            $tempVersion = [Version]$Version
-            if( $TempVersion -and ($TempVersion.Build -lt 0) )
+            if( $module.Version -ge $MinimumVersion -and $module.Version -le $MaximumVersion )
             {
-                $Version = [Version]('{0}.{1}.0' -f $TempVersion.Major, $TempVersion.Minor)
+                Write-WhiskeyDebug "Module $($Name) $($module.Version) already imported."
+                return
+            }
+
+            Write-WhiskeyDebug "Removing module $($Name) $($module.Version)."
+            $module | Remove-Module -Force
+        }
+        
+        $importFailed = $true
+        try
+        {
+            Import-Module -Name $Name `
+                          -MinimumVersion $MinimumVersion `
+                          -MaximumVersion $MaximumVersion `
+                          -Global
+            Write-WhiskeyDebug "Imported module $($Name) $((Get-Module -Name $Name).Version)."
+            $importFailed = $false
+        }
+        finally
+        {
+            if( $importFailed )
+            {
+                Write-WhiskeyDebug "Exception importing $($Name) module: $($Global:Error[0])"
+
+                Write-WhiskeyDebug '\PSModulePath\' -Indent
+                $env:PSModulePath -split ([IO.Path]::PathSeparator) | Write-WhiskeyDebug
+                Write-WhiskeyDebug '/PSModulePath/' -Outdent
+
+                Write-WhiskeyDebug '\Get-Module\' -Indent
+                Get-Module | Out-String | Write-WhiskeyDebug
+                Write-WhiskeyDebug '/Get-Module/' -Outdent
+
+                Write-WhiskeyDebug '\Get-Module -ListAvailable\' -Indent
+                Get-Module -ListAvailable | Out-String | Write-WhiskeyDebug
+                Write-WhiskeyDebug '/Get-Module -ListAvailable/' -Outdent
             }
         }
-
-        $module = 
-            Find-Module -Name $Name -AllVersions -AllowPrerelease:$AllowPrerelease |
-            Where-Object { $_.Version.ToString() -like $Version } |
-            Sort-Object -Property 'Version' -Descending
     }
-    else
+
+    try
     {
-        $atVersionString = ''
-        $module = Find-Module -Name $Name -AllowPrerelease:$AllowPrerelease -ErrorAction Ignore
-    }
+        Register-WhiskeyPSModulePath -PSModulesRoot $BuildRoot
 
-    if( -not $module )
+        # If you want to upgrade the PackageManagement and PowerShellGet versions, you must also update:
+        # * Install-WhiskeyRequiredModule
+        Import -Name 'PackageManagement' `
+               -MinimumVersion $script:pkgMgmtMinVersion `
+               -MaximumVersion $script:pkgMgmtMaxVersion
+        Import -Name 'PowerShellGet' -MinimumVersion $script:psGetMinVersion -MaximumVersion $script:psGetMaxVersion
+
+        $allowPrereleaseArg = Get-AllowPrereleaseArg -CommandName 'Find-Module' -AllowPrerelease:$AllowPrerelease
+
+        Write-WhiskeyDebug -Message ('{0}  {1} ->' -f $Name,$Version)
+        if( $Version )
+        {
+            $atVersionString = ' at version {0}' -f $Version
+
+            if( -not [Management.Automation.WildcardPattern]::ContainsWildcardCharacters($version) -and `
+                [Version]::TryParse($Version,[ref]$null) )
+            {
+                $tempVersion = [Version]$Version
+                if( $tempVersion -and ($tempVersion.Build -lt 0) )
+                {
+                    $Version = [Version]('{0}.{1}.0' -f $tempVersion.Major, $tempVersion.Minor)
+                }
+            }
+
+            Write-WhiskeyDebug -Message "Searching for module $($Name) $($Version)."
+            $module = 
+                Find-Module -Name $Name -AllVersions @allowPrereleaseArg |
+                Where-Object { $_.Version.ToString() -like $Version } |
+                Sort-Object -Property 'Version' -Descending
+        }
+        else
+        {
+            Write-WhiskeyDebug -Message "Searching for latest version of module $($Name)."
+            $atVersionString = ''
+            $module = Find-Module -Name $Name @allowPrereleaseArg -ErrorAction Ignore
+        }
+
+        if( -not $module )
+        {
+            $registeredRepositories = Get-PSRepository | ForEach-Object { ('{0} ({1})' -f $_.Name,$_.SourceLocation) }
+            $registeredRepositories = $registeredRepositories -join ('{0} * ' -f [Environment]::NewLine)
+            Write-WhiskeyError -Message ('Failed to find PowerShell module {0}{1} in any of the registered PowerShell repositories:{2} {2} * {3} {2}' -f $Name, $atVersionString, [Environment]::NewLine, $registeredRepositories)
+            return
+        }
+
+        $module = $module | Select-Object -First 1
+        Write-WhiskeyDebug ($module | Format-List | Out-String)
+        Write-WhiskeyDebug -Message ('{0}  {1}    {2}' -f (' ' * $Name.Length),(' ' * $Version.Length),$module.Version)
+        return $module
+    }
+    finally
     {
-        $registeredRepositories = Get-PSRepository | ForEach-Object { ('{0} ({1})' -f $_.Name,$_.SourceLocation) }
-        $registeredRepositories = $registeredRepositories -join ('{0} * ' -f [Environment]::NewLine)
-        Write-WhiskeyError -Message ('Failed to find PowerShell module {0}{1} in any of the registered PowerShell repositories:{2} {2} * {3} {2}' -f $Name, $atVersionString, [Environment]::NewLine, $registeredRepositories)
-        return
+        Write-WhiskeyDebug '/Find-WhiskeyPowerShellModule/' -Outdent
     }
-
-    $module = $module | Select-Object -First 1
-    Write-WhiskeyDebug -Message ('{0}  {1}    {2}' -f (' ' * $Name.Length),(' ' * $Version.Length),$module.Version)
-    return $module
 }